update GroupedCircularBoxplot.R
Modified by Ranjan: Allow for more flexibility in controlling the plotting character, the margins, and also made the line widths relative. One thing might be worth controlling is the position where the axis is on the outside of a plot with many groups. The center point is made optional since if detracts from many groups plots.